fix: moved v2 schemas into a single file
