+++ src/licenses/v2/licenses-v2.test.ts
@@ -1,10 +1,10 @@
-import { Capability, License, PaginatedLicenses, Plan } from '../../types';
 import prismaClient from '../../../test-utils/prisma/prisma-client';
 import { testUuids } from '../../../test-utils/scripts/create-salable-test-data';
 import getEndTime from '../../../test-utils/helpers/get-end-time';
 import { initSalable } from '../../index';
 import { randomUUID } from 'crypto';
 import { addMonths } from 'date-fns';
+import { LicenseSchema, PaginatedLicensesSchema, PlanSchema } from '../../schemas/v2/schemas-v2';
 
 const licenseUuid = randomUUID();
 const licenseTwoUuid = randomUUID();
@@ -29,21 +29,21 @@ describe('Licenses V2 Tests', () => {
   it('getOne: Should successfully fetch the specified license', async () => {
     const data = await salable.licenses.getOne(licenseUuid);
 
-    expect(data).toEqual(licenseSchema);
+    expect(data).toEqual(LicenseSchema);
     expect(data).not.toHaveProperty('plan');
   });
 
   it('getOne (w/ search params): Should successfully fetch the specified license', async () => {
     const dataWithSearchParams = await salable.licenses.getOne(licenseUuid, { expand: ['plan'] });
 
-    expect(dataWithSearchParams).toEqual({ ...licenseSchema, plan: planSchema });
-    expect(dataWithSearchParams).toHaveProperty('plan', planSchema);
+    expect(dataWithSearchParams).toEqual({ ...LicenseSchema, plan: PlanSchema });
+    expect(dataWithSearchParams).toHaveProperty('plan', PlanSchema);
   });
 
   it('getAll: Should successfully fetch licenses', async () => {
     const data = await salable.licenses.getAll();
 
-    expect(data).toEqual(paginatedLicensesSchema);
+    expect(data).toEqual(PaginatedLicensesSchema);
   });
 
   it('getAll (w/ search params): Should successfully fetch licenses', async () => {
@@ -56,7 +56,7 @@ describe('Licenses V2 Tests', () => {
     expect(dataWithSearchParams).toEqual({
       first: expect.any(String),
       last: expect.any(String),
-      data: expect.arrayContaining([licenseSchema]),
+      data: expect.arrayContaining([LicenseSchema]),
     });
     expect(dataWithSearchParams.data.length).toEqual(3);
     for (const license of dataWithSearchParams.data) {
@@ -90,7 +90,7 @@ describe('Licenses V2 Tests', () => {
   it('getForPurchaser: Should successfully fetch a purchasers licenses', async () => {
     const data = await salable.licenses.getForPurchaser({ purchaser: testPurchaser, productUuid: testUuids.productUuid });
 
-    expect(data).toEqual(expect.arrayContaining([expect.objectContaining(licenseSchema)]));
+    expect(data).toEqual(expect.arrayContaining([expect.objectContaining(LicenseSchema)]));
   });
 
   it('getForPurchaser (w/ search params): Should successfully fetch a purchasers licenses', async () => {
@@ -103,7 +103,7 @@ describe('Licenses V2 Tests', () => {
     expect(dataWithSearchParams).toEqual(
       expect.arrayContaining([
         expect.objectContaining({
-          ...licenseSchema,
+          ...LicenseSchema,
           status: 'ACTIVE',
           purchaser: testPurchaser,
           productUuid: testUuids.productUuid,
@@ -115,7 +115,7 @@ describe('Licenses V2 Tests', () => {
   it('getForGranteeId: Should successfully fetch a grantees licenses', async () => {
     const data = await salable.licenses.getForGranteeId(testGrantee);
 
-    expect(data).toEqual(expect.arrayContaining([expect.objectContaining(licenseSchema)]));
+    expect(data).toEqual(expect.arrayContaining([expect.objectContaining(LicenseSchema)]));
   });
 
   it('getForGranteeId (w/ search params): Should successfully fetch a grantees licenses', async () => {
@@ -124,8 +124,8 @@ describe('Licenses V2 Tests', () => {
     expect(dataWithSearchParams).toEqual(
       expect.arrayContaining([
         expect.objectContaining({
-          ...licenseSchema,
-          plan: planSchema,
+          ...LicenseSchema,
+          plan: PlanSchema,
         }),
       ]),
     );
@@ -140,7 +140,7 @@ describe('Licenses V2 Tests', () => {
       endTime: '2025-07-06T12:00:00.000Z',
     });
 
-    expect(data).toEqual(expect.objectContaining(licenseSchema));
+    expect(data).toEqual(expect.objectContaining(LicenseSchema));
   });
 
   it('createMany: Should successfully create multiple licenses', async () => {
@@ -162,7 +162,7 @@ describe('Licenses V2 Tests', () => {
     ]);
 
     expect(data.length).toEqual(2);
-    expect(data).toEqual(expect.arrayContaining([expect.objectContaining(licenseSchema)]));
+    expect(data).toEqual(expect.arrayContaining([expect.objectContaining(LicenseSchema)]));
   });
 
   it('update: Should successfully update a license', async () => {
@@ -187,7 +187,7 @@ describe('Licenses V2 Tests', () => {
     expect(data).toEqual(
       expect.arrayContaining([
         expect.objectContaining({
-          ...licenseSchema,
+          ...LicenseSchema,
           granteeId: 'updated-grantee-id',
         }),
       ]),
@@ -234,74 +234,6 @@ describe('Licenses V2 Tests', () => {
     expect(data).toBeUndefined();
   });
 });
-
-const licenseCapabilitySchema: Capability = {
-  uuid: expect.any(String),
-  productUuid: expect.any(String),
-  name: expect.any(String),
-  status: expect.any(String),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  updatedAt: expect.any(String),
-};
-
-const licenseSchema: License = {
-  uuid: expect.any(String),
-  name: expect.toBeOneOf([expect.any(String), null]),
-  email: expect.toBeOneOf([expect.any(String), null]),
-  subscriptionUuid: expect.toBeOneOf([expect.any(String), null]),
-  status: expect.toBeOneOf(['ACTIVE', 'CANCELED', 'EVALUATION', 'SCHEDULED', 'TRIALING', 'INACTIVE']),
-  granteeId: expect.toBeOneOf([expect.any(String), null]),
-  paymentService: expect.toBeOneOf(['ad-hoc', 'salable', 'stripe_existing']),
-  purchaser: expect.any(String),
-  type: expect.toBeOneOf(['licensed', 'metered', 'perSeat', 'customId', 'user']),
-  productUuid: expect.any(String),
-  planUuid: expect.any(String),
-  capabilities: expect.arrayContaining([licenseCapabilitySchema]),
-  metadata: expect.toBeOneOf([expect.anything(), null]),
-  startTime: expect.any(String),
-  endTime: expect.any(String),
-  updatedAt: expect.any(String),
-  isTest: expect.any(Boolean),
-  cancelAtPeriodEnd: expect.any(Boolean),
-};
-
-const paginatedLicensesSchema: PaginatedLicenses = {
-  first: expect.toBeOneOf([expect.any(String), null]),
-  last: expect.toBeOneOf([expect.any(String), null]),
-  data: expect.arrayContaining([licenseSchema]),
-};
-
-const planSchema: Plan = {
-  uuid: expect.any(String),
-  name: expect.any(String),
-  slug: expect.any(String),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  displayName: expect.any(String),
-  status: expect.any(String),
-  trialDays: expect.toBeOneOf([expect.any(Number), null]),
-  evaluation: expect.any(Boolean),
-  evalDays: expect.any(Number),
-  perSeatAmount: expect.any(Number),
-  maxSeatAmount: expect.any(Number),
-  organisation: expect.any(String),
-  visibility: expect.any(String),
-  licenseType: expect.any(String),
-  hasAcceptedTransaction: expect.any(Boolean),
-  interval: expect.any(String),
-  length: expect.any(Number),
-  active: expect.any(Boolean),
-  planType: expect.any(String),
-  pricingType: expect.any(String),
-  environment: expect.any(String),
-  isTest: expect.any(Boolean),
-  paddlePlanId: expect.toBeOneOf([expect.any(String), null]),
-  productUuid: expect.any(String),
-  salablePlan: expect.any(Boolean),
-  type: expect.toBeOneOf([expect.any(String), undefined]),
-  updatedAt: expect.any(String),
-  archivedAt: expect.toBeOneOf([expect.any(String), null]),
-  features: expect.toBeOneOf([expect.anything(), undefined]),
-};
 
 const generateTestData = async () => {
   await prismaClient.subscription.create({
