+++ src/plans/v2/plan-v2.test.ts
@@ -1,6 +1,6 @@
-import { Plan, PlanCapability, PlanCheckout, PlanCurrency, PlanFeature } from '../../types';
 import { testUuids } from '../../../test-utils/scripts/create-salable-test-data';
 import { initSalable } from '../../index';
+import { PlanCapabilitySchema, PlanCheckoutLinkSchema, PlanCurrencySchema, PlanFeatureSchema, PlanSchema } from '../../schemas/v2/schemas-v2';
 
 describe('Plans V2 Tests', () => {
   const apiKey = testUuids.devApiKeyV2;
@@ -10,13 +10,13 @@ describe('Plans V2 Tests', () => {
 
   it('getOne: should successfully fetch all products', async () => {
     const data = await salable.plans.getOne(planUuid);
-    expect(data).toEqual(planSchema);
+    expect(data).toEqual(PlanSchema);
   });
 
   it('getOne (w / search params): should successfully fetch a plan', async () => {
     const data = await salable.plans.getOne(planUuid, { expand: ['capabilities', 'capabilities.capability', 'features', 'features.feature', 'features.enumValue', 'currencies', 'currencies.currency'] });
 
-    expect(data).toEqual(planSchema);
+    expect(data).toEqual(PlanSchema);
   });
 
   it('getCheckoutLink (w / required params): should successfully fetch checkout link for plan', async () => {
@@ -65,75 +65,3 @@ describe('Plans V2 Tests', () => {
     expect(data).toEqual(expect.arrayContaining([PlanCurrencySchema]));
   });
 });
-
-const planSchema: Plan = {
-  uuid: expect.any(String),
-  name: expect.any(String),
-  slug: expect.any(String),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  displayName: expect.any(String),
-  status: expect.any(String),
-  trialDays: expect.toBeOneOf([expect.any(Number), null]),
-  evaluation: expect.any(Boolean),
-  evalDays: expect.any(Number),
-  perSeatAmount: expect.any(Number),
-  maxSeatAmount: expect.any(Number),
-  organisation: expect.any(String),
-  visibility: expect.any(String),
-  licenseType: expect.any(String),
-  hasAcceptedTransaction: expect.any(Boolean),
-  interval: expect.any(String),
-  length: expect.any(Number),
-  active: expect.any(Boolean),
-  planType: expect.any(String),
-  pricingType: expect.any(String),
-  environment: expect.any(String),
-  isTest: expect.any(Boolean),
-  paddlePlanId: expect.toBeOneOf([expect.any(String), null]),
-  productUuid: expect.any(String),
-  salablePlan: expect.any(Boolean),
-  type: expect.toBeOneOf([expect.any(String), undefined]),
-  updatedAt: expect.any(String),
-  archivedAt: expect.toBeOneOf([expect.any(String), null]),
-  features: expect.toBeOneOf([expect.anything(), undefined]),
-};
-
-const PlanCheckoutLinkSchema: PlanCheckout = {
-  checkoutUrl: expect.any(String),
-};
-
-const PlanFeatureSchema: PlanFeature = {
-  enumValue: expect.toBeOneOf([expect.anything(), null]),
-  enumValueUuid: expect.any(String),
-  feature: expect.toBeOneOf([expect.anything()]),
-  featureUuid: expect.any(String),
-  isUnlimited: expect.any(Boolean),
-  isUsage: expect.any(Boolean),
-  maxUsage: expect.any(Number),
-  minUsage: expect.any(Number),
-  planUuid: expect.any(String),
-  pricePerUnit: expect.any(Number),
-  updatedAt: expect.any(String),
-  value: expect.any(String),
-};
-
-const PlanCapabilitySchema: PlanCapability = {
-  planUuid: expect.any(String),
-  capabilityUuid: expect.any(String),
-  updatedAt: expect.any(String),
-  capability: expect.toBeOneOf([expect.anything()]),
-};
-
-const PlanCurrencySchema: PlanCurrency = {
-  planUuid: expect.any(String),
-  currencyUuid: expect.any(String),
-  price: expect.any(Number),
-  paymentIntegrationPlanId: expect.any(String),
-  hasAcceptedTransaction: expect.any(Boolean),
-  currency: {
-    uuid: expect.any(String),
-    shortName: expect.any(String),
-    longName: expect.any(String),
-    symbol: expect.any(String),
-  },
-};
