+++ src/products/v2/product-v2.test.ts
@@ -1,6 +1,6 @@
-import { Plan, Product, ProductCapability, ProductCurrency, ProductFeature, ProductPricingTable, Version } from '../../types';
 import { testUuids } from '../../../test-utils/scripts/create-salable-test-data';
 import { initSalable } from '../../index';
+import { ProductCapabilitySchema, ProductCurrencySchema, ProductFeatureSchema, ProductPlanSchema, ProductPricingTableSchema, ProductSchema } from '../../schemas/v2/schemas-v2';
 
 describe('Products V2 Tests', () => {
   const apiKey = testUuids.devApiKeyV2;
@@ -73,95 +73,3 @@ describe('Products V2 Tests', () => {
     expect(data).toEqual(expect.arrayContaining([ProductCurrencySchema]));
   });
 });
-
-const ProductSchema: Product = {
-  uuid: expect.any(String),
-  name: expect.any(String),
-  slug: expect.any(String),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  logoUrl: expect.toBeOneOf([expect.any(String), null]),
-  displayName: expect.toBeOneOf([expect.any(String), null]),
-  organisation: expect.any(String),
-  status: expect.any(String),
-  paid: expect.any(Boolean),
-  isTest: expect.any(Boolean),
-  organisationPaymentIntegrationUuid: expect.any(String),
-  paymentIntegrationProductId: expect.toBeOneOf([expect.any(String), null]),
-  appType: expect.any(String),
-  updatedAt: expect.any(String),
-  archivedAt: expect.toBeOneOf([expect.any(String), null]),
-};
-
-const ProductPricingTableSchema: ProductPricingTable = {
-  ...ProductSchema,
-  features: expect.toBeOneOf([expect.anything(), undefined]),
-  currencies: expect.toBeOneOf([expect.anything(), undefined]),
-  plans: expect.toBeOneOf([expect.anything(), undefined]),
-  status: expect.any(String),
-  updatedAt: expect.any(String),
-  uuid: expect.any(String),
-};
-const ProductPlanSchema: Plan = {
-  uuid: expect.any(String),
-  name: expect.any(String),
-  slug: expect.any(String),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  displayName: expect.any(String),
-  status: expect.any(String),
-  trialDays: expect.toBeOneOf([expect.any(Number), null]),
-  evaluation: expect.any(Boolean),
-  evalDays: expect.any(Number),
-  perSeatAmount: expect.any(Number),
-  maxSeatAmount: expect.any(Number),
-  organisation: expect.any(String),
-  visibility: expect.any(String),
-  licenseType: expect.any(String),
-  hasAcceptedTransaction: expect.any(Boolean),
-  interval: expect.any(String),
-  length: expect.any(Number),
-  active: expect.any(Boolean),
-  planType: expect.any(String),
-  pricingType: expect.any(String),
-  environment: expect.any(String),
-  isTest: expect.any(Boolean),
-  paddlePlanId: expect.toBeOneOf([expect.any(String), null]),
-  productUuid: expect.any(String),
-  salablePlan: expect.any(Boolean),
-  type: expect.toBeOneOf([expect.any(String), undefined]),
-  updatedAt: expect.any(String),
-  features: expect.toBeOneOf([expect.anything(), undefined]),
-  archivedAt: expect.toBeOneOf([expect.any(String), null]),
-};
-
-const ProductFeatureSchema: ProductFeature = {
-  defaultValue: expect.any(String),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  displayName: expect.any(String),
-  featureEnumOptions: expect.toBeOneOf([expect.anything()]),
-  name: expect.any(String),
-  productUuid: expect.any(String),
-  showUnlimited: expect.any(Boolean),
-  sortOrder: expect.any(Number),
-  status: expect.any(String),
-  updatedAt: expect.any(String),
-  uuid: expect.any(String),
-  valueType: expect.any(String),
-  variableName: expect.any(String),
-  visibility: expect.any(String),
-};
-
-const ProductCapabilitySchema: ProductCapability = {
-  uuid: expect.any(String),
-  name: expect.any(String),
-  description: expect.any(String),
-  status: expect.any(String),
-  productUuid: expect.any(String),
-  updatedAt: expect.any(String),
-};
-
-const ProductCurrencySchema: ProductCurrency = {
-  productUuid: expect.any(String),
-  currencyUuid: expect.any(String),
-  defaultCurrency: expect.any(Boolean),
-  currency: expect.toBeOneOf([expect.anything()]),
-};
