+++ src/schemas/v2/schemas-v2.ts
@@ -1,0 +1,452 @@
+import {
+  EnumValue,
+  Event,
+  EventTypeEnum,
+  Invoice,
+  PaginatedLicenses,
+  PaginatedSubscription,
+  PaginatedSubscriptionInvoice,
+  PaginatedUsageRecords,
+  Plan,
+  PlanCapability,
+  PlanCheckout,
+  PlanCurrency,
+  PlanFeature,
+  PricingTable,
+  Product,
+  ProductCapability,
+  ProductCurrency,
+  ProductFeature,
+  ProductPricingTable,
+  Session,
+  Subscription,
+  UsageRecord,
+} from '../../types';
+import { Capability, EventStatus, License } from '@prisma/client';
+
+export const EventSchema: Event = {
+  uuid: expect.any(String),
+  type: expect.toBeOneOf(Object.values(EventTypeEnum)) as EventTypeEnum,
+  organisation: expect.any(String),
+  status: expect.toBeOneOf(Object.values(EventStatus)) as EventStatus,
+  isTest: expect.any(Boolean),
+  retries: expect.any(Number),
+  errorMessage: expect.toBeOneOf([expect.any(String), null]),
+  errorCode: expect.toBeOneOf([expect.any(String), null]),
+  createdAt: expect.any(String),
+  updatedAt: expect.any(String),
+};
+
+export const LicenseCapabilitySchema: Capability = {
+  uuid: expect.any(String),
+  productUuid: expect.any(String),
+  name: expect.any(String),
+  status: expect.any(String),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  updatedAt: expect.any(String),
+};
+
+export const LicenseSchema: License = {
+  uuid: expect.any(String),
+  name: expect.toBeOneOf([expect.any(String), null]),
+  email: expect.toBeOneOf([expect.any(String), null]),
+  subscriptionUuid: expect.toBeOneOf([expect.any(String), null]),
+  status: expect.toBeOneOf(['ACTIVE', 'CANCELED', 'EVALUATION', 'SCHEDULED', 'TRIALING', 'INACTIVE']),
+  granteeId: expect.toBeOneOf([expect.any(String), null]),
+  paymentService: expect.toBeOneOf(['ad-hoc', 'salable', 'stripe_existing']),
+  purchaser: expect.any(String),
+  type: expect.toBeOneOf(['licensed', 'metered', 'perSeat', 'customId', 'user']),
+  productUuid: expect.any(String),
+  planUuid: expect.any(String),
+  capabilities: expect.arrayContaining([LicenseCapabilitySchema]),
+  metadata: expect.toBeOneOf([expect.anything(), null]),
+  startTime: expect.any(String),
+  endTime: expect.any(String),
+  updatedAt: expect.any(String),
+  isTest: expect.any(Boolean),
+  cancelAtPeriodEnd: expect.any(Boolean),
+};
+
+export const PaginatedLicensesSchema: PaginatedLicenses = {
+  first: expect.toBeOneOf([expect.any(String), null]),
+  last: expect.toBeOneOf([expect.any(String), null]),
+  data: expect.arrayContaining([LicenseSchema]),
+};
+
+export const PlanSchema: Plan = {
+  uuid: expect.any(String),
+  name: expect.any(String),
+  slug: expect.any(String),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  displayName: expect.any(String),
+  status: expect.any(String),
+  trialDays: expect.toBeOneOf([expect.any(Number), null]),
+  evaluation: expect.any(Boolean),
+  evalDays: expect.any(Number),
+  perSeatAmount: expect.any(Number),
+  maxSeatAmount: expect.any(Number),
+  organisation: expect.any(String),
+  visibility: expect.any(String),
+  licenseType: expect.any(String),
+  hasAcceptedTransaction: expect.any(Boolean),
+  interval: expect.any(String),
+  length: expect.any(Number),
+  active: expect.any(Boolean),
+  planType: expect.any(String),
+  pricingType: expect.any(String),
+  environment: expect.any(String),
+  isTest: expect.any(Boolean),
+  paddlePlanId: expect.toBeOneOf([expect.any(String), null]),
+  productUuid: expect.any(String),
+  salablePlan: expect.any(Boolean),
+  type: expect.toBeOneOf([expect.any(String), undefined]),
+  updatedAt: expect.any(String),
+  features: expect.toBeOneOf([expect.anything(), undefined]),
+  currencies: expect.toBeOneOf([expect.anything(), undefined]),
+  archivedAt: expect.toBeOneOf([expect.any(String), null]),
+};
+
+export const PlanCheckoutLinkSchema: PlanCheckout = {
+  checkoutUrl: expect.any(String),
+};
+
+export const PlanFeatureSchema: PlanFeature = {
+  enumValue: expect.toBeOneOf([expect.anything(), null]),
+  enumValueUuid: expect.any(String),
+  feature: expect.toBeOneOf([expect.anything()]),
+  featureUuid: expect.any(String),
+  isUnlimited: expect.any(Boolean),
+  isUsage: expect.any(Boolean),
+  maxUsage: expect.any(Number),
+  minUsage: expect.any(Number),
+  planUuid: expect.any(String),
+  pricePerUnit: expect.any(Number),
+  updatedAt: expect.any(String),
+  value: expect.any(String),
+};
+
+export const PlanCapabilitySchema: PlanCapability = {
+  planUuid: expect.any(String),
+  capabilityUuid: expect.any(String),
+  updatedAt: expect.any(String),
+  capability: expect.toBeOneOf([expect.anything()]),
+};
+
+export const PlanCurrencySchema: PlanCurrency = {
+  planUuid: expect.any(String),
+  currencyUuid: expect.any(String),
+  price: expect.any(Number),
+  paymentIntegrationPlanId: expect.any(String),
+  hasAcceptedTransaction: expect.any(Boolean),
+  currency: {
+    uuid: expect.any(String),
+    shortName: expect.any(String),
+    longName: expect.any(String),
+    symbol: expect.any(String),
+  },
+};
+
+export const PricingTableSchema: PricingTable = {
+  customTheme: expect.toBeOneOf([expect.any(String), null]),
+  productUuid: expect.any(String),
+  featuredPlanUuid: expect.toBeOneOf([expect.any(String), null]),
+  name: expect.any(String),
+  status: expect.any(String),
+  theme: expect.any(String),
+  text: expect.toBeOneOf([expect.any(String), null]),
+  title: expect.toBeOneOf([expect.any(String), null]),
+  updatedAt: expect.any(String),
+  uuid: expect.any(String),
+  featureOrder: expect.any(String),
+  features: expect.toBeOneOf([expect.anything(), undefined]),
+  product: expect.toBeOneOf([expect.anything(), undefined]),
+  plans: expect.toBeOneOf([expect.anything(), undefined]),
+};
+
+export const ProductSchema: Product = {
+  uuid: expect.any(String),
+  name: expect.any(String),
+  slug: expect.any(String),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  logoUrl: expect.toBeOneOf([expect.any(String), null]),
+  displayName: expect.toBeOneOf([expect.any(String), null]),
+  organisation: expect.any(String),
+  status: expect.any(String),
+  paid: expect.any(Boolean),
+  isTest: expect.any(Boolean),
+  organisationPaymentIntegrationUuid: expect.any(String),
+  paymentIntegrationProductId: expect.toBeOneOf([expect.any(String), null]),
+  appType: expect.any(String),
+  updatedAt: expect.any(String),
+  archivedAt: expect.toBeOneOf([expect.any(String), null]),
+};
+
+export const ProductPricingTableSchema: ProductPricingTable = {
+  ...ProductSchema,
+  features: expect.toBeOneOf([expect.anything(), undefined]),
+  currencies: expect.toBeOneOf([expect.anything(), undefined]),
+  plans: expect.toBeOneOf([expect.anything(), undefined]),
+  status: expect.any(String),
+  updatedAt: expect.any(String),
+  uuid: expect.any(String),
+};
+export const ProductPlanSchema: Plan = {
+  uuid: expect.any(String),
+  name: expect.any(String),
+  slug: expect.any(String),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  displayName: expect.any(String),
+  status: expect.any(String),
+  trialDays: expect.toBeOneOf([expect.any(Number), null]),
+  evaluation: expect.any(Boolean),
+  evalDays: expect.any(Number),
+  perSeatAmount: expect.any(Number),
+  maxSeatAmount: expect.any(Number),
+  organisation: expect.any(String),
+  visibility: expect.any(String),
+  licenseType: expect.any(String),
+  hasAcceptedTransaction: expect.any(Boolean),
+  interval: expect.any(String),
+  length: expect.any(Number),
+  active: expect.any(Boolean),
+  planType: expect.any(String),
+  pricingType: expect.any(String),
+  environment: expect.any(String),
+  isTest: expect.any(Boolean),
+  paddlePlanId: expect.toBeOneOf([expect.any(String), null]),
+  productUuid: expect.any(String),
+  salablePlan: expect.any(Boolean),
+  type: expect.toBeOneOf([expect.any(String), undefined]),
+  updatedAt: expect.any(String),
+  features: expect.toBeOneOf([expect.anything(), undefined]),
+  archivedAt: expect.toBeOneOf([expect.any(String), null]),
+};
+
+export const ProductFeatureSchema: ProductFeature = {
+  defaultValue: expect.any(String),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  displayName: expect.any(String),
+  featureEnumOptions: expect.toBeOneOf([expect.anything()]),
+  name: expect.any(String),
+  productUuid: expect.any(String),
+  showUnlimited: expect.any(Boolean),
+  sortOrder: expect.any(Number),
+  status: expect.any(String),
+  updatedAt: expect.any(String),
+  uuid: expect.any(String),
+  valueType: expect.any(String),
+  variableName: expect.any(String),
+  visibility: expect.any(String),
+};
+
+export const ProductCapabilitySchema: ProductCapability = {
+  uuid: expect.any(String),
+  name: expect.any(String),
+  description: expect.any(String),
+  status: expect.any(String),
+  productUuid: expect.any(String),
+  updatedAt: expect.any(String),
+};
+
+export const ProductCurrencySchema: ProductCurrency = {
+  productUuid: expect.any(String),
+  currencyUuid: expect.any(String),
+  defaultCurrency: expect.any(Boolean),
+  currency: expect.toBeOneOf([expect.anything()]),
+};
+
+export const SubscriptionSchema: Subscription = {
+  uuid: expect.any(String),
+  paymentIntegrationSubscriptionId: expect.any(String),
+  productUuid: expect.any(String),
+  type: expect.any(String),
+  isTest: expect.any(Boolean),
+  cancelAtPeriodEnd: expect.any(Boolean),
+  email: expect.toBeOneOf([expect.any(String), null]),
+  owner: expect.toBeOneOf([expect.any(String), null]),
+  organisation: expect.any(String),
+  quantity: expect.any(Number),
+  status: expect.any(String),
+  createdAt: expect.any(String),
+  updatedAt: expect.any(String),
+  expiryDate: expect.any(String),
+  lineItemIds: expect.toBeOneOf([expect.toBeArray(), null]),
+  planUuid: expect.any(String),
+};
+
+export const EnumValueSchema: EnumValue = {
+  uuid: expect.any(String),
+  name: expect.any(String),
+  featureUuid: expect.any(String),
+  updatedAt: expect.any(String),
+}
+
+export const SessionSchema: Session = {
+  sessionToken: expect.any(String),
+};
+
+export const PaginationSubscriptionSchema: PaginatedSubscription = {
+  first: expect.any(String),
+  last: expect.any(String),
+  data: expect.arrayContaining([SubscriptionSchema]),
+};
+
+export const InvoiceSchema: Invoice = {
+  id: expect.any(String),
+  object: expect.any(String),
+  account_country: expect.any(String),
+  account_name: expect.any(String),
+  account_tax_ids: expect.toBeOneOf([expect.toBeArray(), null]),
+  amount_due: expect.any(Number),
+  amount_paid: expect.any(Number),
+  amount_overpaid: expect.any(Number),
+  amount_remaining: expect.any(Number),
+  amount_shipping: expect.any(Number),
+  application: expect.toBeOneOf([expect.any(String), null]),
+  application_fee_amount: expect.toBeOneOf([expect.any(Number), null]),
+  attempt_count: expect.any(Number),
+  attempted: expect.any(Boolean),
+  auto_advance: expect.any(Boolean),
+  automatic_tax: expect.toBeObject(),
+  automatically_finalizes_at: expect.toBeOneOf([expect.any(Number), null]),
+  billing_reason: expect.any(String),
+  charge: expect.any(String),
+  collection_method: expect.any(String),
+  created: expect.any(Number),
+  currency: expect.any(String),
+  custom_fields: expect.toBeOneOf([expect.toBeArray(), null]),
+  customer: expect.any(String),
+  customer_address: expect.toBeOneOf([expect.toBeObject(), null]),
+  customer_email: expect.any(String),
+  customer_name: expect.toBeOneOf([expect.any(String), null]),
+  customer_phone: expect.toBeOneOf([expect.any(String), null]),
+  customer_shipping: expect.toBeOneOf([expect.toBeObject, null]),
+  customer_tax_exempt: expect.any(String),
+  customer_tax_ids: expect.toBeOneOf([expect.toBeArray(), null]),
+  default_payment_method: expect.toBeOneOf([expect.any(String), null]),
+  default_source: expect.toBeOneOf([expect.any(String), null]),
+  default_tax_rates: expect.toBeOneOf([expect.toBeArray(), null]),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  discount: expect.toBeOneOf([expect.toBeObject(), null]),
+  discounts: expect.toBeOneOf([expect.toBeArray(), null]),
+  due_date: expect.toBeOneOf([expect.any(Number), null]),
+  effective_at: expect.any(Number),
+  ending_balance: expect.any(Number),
+  footer: expect.toBeOneOf([expect.any(String), null]),
+  from_invoice: expect.toBeOneOf([expect.toBeObject(), null]),
+  hosted_invoice_url: expect.any(String),
+  invoice_pdf: expect.any(String),
+  issuer: expect.toBeObject(),
+  last_finalization_error: expect.toBeOneOf([expect.toBeObject(), null]),
+  latest_revision: expect.toBeOneOf([expect.any(String), null]),
+  lines: expect.toBeObject(),
+  livemode: expect.any(Boolean),
+  metadata: expect.toBeObject(),
+  next_payment_attempt: expect.toBeOneOf([expect.any(Number), null]),
+  number: expect.any(String),
+  on_behalf_of: expect.toBeOneOf([expect.any(String), null]),
+  paid: expect.any(Boolean),
+  paid_out_of_band: expect.any(Boolean),
+  parent: expect.toBeObject(),
+  payment_intent: expect.any(String),
+  payment_settings: expect.toBeObject(),
+  period_end: expect.any(Number),
+  period_start: expect.any(Number),
+  post_payment_credit_notes_amount: expect.any(Number),
+  pre_payment_credit_notes_amount: expect.any(Number),
+  quote: expect.toBeOneOf([expect.any(String), null]),
+  receipt_number: expect.toBeOneOf([expect.any(String), null]),
+  rendering: expect.toBeOneOf([expect.toBeObject(), null]),
+  rendering_options: expect.toBeOneOf([expect.toBeObject(), undefined]),
+  shipping_cost: expect.toBeOneOf([expect.toBeObject(), null]),
+  shipping_details: expect.toBeOneOf([expect.toBeObject(), null]),
+  starting_balance: expect.any(Number),
+  statement_descriptor: expect.toBeOneOf([expect.any(String), null]),
+  status: expect.any(String),
+  status_transitions: expect.toBeObject(),
+  subscription: expect.any(String),
+  subscription_details: expect.toBeObject(),
+  subtotal: expect.any(Number),
+  subtotal_excluding_tax: expect.any(Number),
+  tax: expect.toBeOneOf([expect.any(Number), null]),
+  test_clock: expect.toBeOneOf([expect.any(String), null]),
+  total: expect.any(Number),
+  total_discount_amounts: expect.toBeOneOf([expect.toBeArray(), null]),
+  total_excluding_tax: expect.any(Number),
+  total_pretax_credit_amounts: expect.toBeOneOf([expect.toBeArray(), null]),
+  total_tax_amounts: expect.toBeArray(),
+  total_taxes: expect.toBeArray(),
+  transfer_data: expect.toBeOneOf([expect.toBeObject(), null]),
+  webhooks_delivered_at: expect.toBeOneOf([expect.any(Number), null]),
+};
+
+export const StripeInvoiceSchema: PaginatedSubscriptionInvoice = {
+  first: expect.any(String),
+  last: expect.any(String),
+  hasMore: expect.any(Boolean),
+  data: [InvoiceSchema],
+};
+
+export const StripePaymentMethodSchema = {
+  id: expect.any(String),
+  object: expect.any(String),
+  allow_redisplay: expect.any(String),
+  billing_details: expect.objectContaining({
+    address: {
+      city: expect.toBeOneOf([expect.any(String), null]),
+      country: expect.toBeOneOf([expect.any(String), null]),
+      line1: expect.toBeOneOf([expect.any(String), null]),
+      line2: expect.toBeOneOf([expect.any(String), null]),
+      postal_code: expect.toBeOneOf([expect.any(String), null]),
+      state: expect.toBeOneOf([expect.any(String), null]),
+    },
+    email: expect.toBeOneOf([expect.any(String), null]),
+    name: expect.toBeOneOf([expect.any(String), null]),
+    phone: expect.toBeOneOf([expect.any(String), null]),
+  }),
+  card: expect.objectContaining({
+    brand: expect.any(String),
+    checks: {
+      address_line1_check: expect.toBeOneOf([expect.any(String), null]),
+      address_postal_code_check: expect.toBeOneOf([expect.any(String), null]),
+      cvc_check: expect.any(String),
+    },
+    country: expect.any(String),
+    display_brand: expect.any(String),
+    exp_month: expect.any(Number),
+    exp_year: expect.any(Number),
+    fingerprint: expect.any(String),
+    funding: expect.any(String),
+    generated_from: expect.toBeOneOf([expect.any(String), null]),
+    last4: expect.any(String),
+    networks: expect.objectContaining({
+      available: expect.toBeArray(),
+      preferred: expect.toBeOneOf([expect.any(String), null]),
+    }),
+    three_d_secure_usage: expect.objectContaining({ supported: expect.any(Boolean) }),
+    wallet: expect.toBeOneOf([expect.any(String), null]),
+  }),
+  created: expect.any(Number),
+  customer: expect.any(String),
+  livemode: expect.any(Boolean),
+  metadata: expect.toBeObject(),
+  type: expect.any(String),
+};
+
+export const UsageRecordSchema: UsageRecord = {
+  uuid: expect.any(String),
+  unitCount: expect.any(Number),
+  type: expect.any(String),
+  recordedAt: expect.toBeOneOf([expect.any(String), null]),
+  resetAt: expect.toBeOneOf([expect.any(String), null]),
+  planUuid: expect.any(String),
+  licenseUuid: expect.any(String),
+  createdAt: expect.any(String),
+  updatedAt: expect.any(String),
+};
+
+export const PaginatedUsageRecordsSchema: PaginatedUsageRecords = {
+  first: expect.toBeOneOf([expect.any(String), null]),
+  last: expect.toBeOneOf([expect.any(String), null]),
+  data: expect.arrayContaining([UsageRecordSchema]),
+};
