+++ src/subscriptions/v3/subscriptions-v3.test.ts
@@ -1,22 +1,14 @@
 import prismaClient from '../../../test-utils/prisma/prisma-client';
 import {
-  PaginatedSubscription,
-  Invoice,
-  PaginatedSubscriptionInvoice,
-  PaginatedLicenses,
   SeatActionType,
 } from '../../types';
 import getEndTime from '../../../test-utils/helpers/get-end-time';
 import { testUuids } from '../../../test-utils/scripts/create-salable-test-data';
 import { randomUUID } from 'crypto';
-import {
-  PlanFeatureSchemaV3,
-  LicenseSchemaV3,
-  PlanCurrencySchema,
-  PlanSchemaV3, SubscriptionSchema
-} from '../../schemas/v3/schemas-v3';
+import { PlanFeatureSchemaV3, PlanCurrencySchema, PlanSchemaV3, PaginatedLicensesSchemaV3 } from '../../schemas/v3/schemas-v3';
 import { addMonths } from 'date-fns';
 import { initSalable } from '../../index';
+import { PaginationSubscriptionSchema, StripeInvoiceSchema, StripePaymentMethodSchema, SubscriptionSchema } from '../../schemas/v2/schemas-v2';
 
 const stripeEnvs = JSON.parse(process.env.stripEnvs || '');
 
@@ -54,7 +46,7 @@ describe('Subscriptions V3 Tests', () => {
 
   it('getAll: Should successfully fetch subscriptions', async () => {
     const data = await salable.subscriptions.getAll();
-    expect(data).toEqual(paginationSubscriptionSchema);
+    expect(data).toEqual(PaginationSubscriptionSchema);
   });
 
   it('getAll (w/ search params): Should successfully fetch subscriptions', async () => {
@@ -118,7 +110,7 @@ describe('Subscriptions V3 Tests', () => {
 
   it("getSeats: Should successfully fetch a subscription's seats", async () => {
     const data = await salable.subscriptions.getSeats(perSeatSubscriptionUuid);
-    expect(data).toEqual(paginatedLicensesSchema);
+    expect(data).toEqual(PaginatedLicensesSchemaV3);
   });
 
   it("getSeatCount: Should successfully fetch a subscription's seat count", async () => {
@@ -150,12 +142,12 @@ describe('Subscriptions V3 Tests', () => {
 
   it('getInvoices: Should successfully fetch a subscriptions invoices', async () => {
     const data = await salable.subscriptions.getInvoices(testUuids.subscriptionWithInvoicesUuid);
-    expect(data).toEqual(stripeInvoiceSchema);
+    expect(data).toEqual(StripeInvoiceSchema);
   });
 
   it('getInvoices (w/ search params): Should successfully fetch a subscriptions invoices', async () => {
     const data = await salable.subscriptions.getInvoices(testUuids.subscriptionWithInvoicesUuid, { take: 1 });
-    expect(data).toEqual(stripeInvoiceSchema);
+    expect(data).toEqual(StripeInvoiceSchema);
     expect(data.data.length).toEqual(1);
   });
 
@@ -176,7 +168,7 @@ describe('Subscriptions V3 Tests', () => {
 
   it('getPaymentMethod: Should successfully fetch a subscriptions payment method', async () => {
     const data = await salable.subscriptions.getPaymentMethod(testUuids.subscriptionWithInvoicesUuid);
-    expect(data).toEqual(expect.objectContaining(stripePaymentMethodSchema));
+    expect(data).toEqual(expect.objectContaining(StripePaymentMethodSchema));
   });
 
   it('changePlan: Should successfully change a subscriptions plan', async () => {
@@ -231,160 +223,6 @@ describe('Subscriptions V3 Tests', () => {
     expect(data).toBeUndefined();
   });
 });
-
-const paginatedLicensesSchema: PaginatedLicenses = {
-  first: expect.toBeOneOf([expect.any(String), null]),
-  last: expect.toBeOneOf([expect.any(String), null]),
-  data: expect.arrayContaining([LicenseSchemaV3]),
-};
-
-const paginationSubscriptionSchema: PaginatedSubscription = {
-  first: expect.any(String),
-  last: expect.any(String),
-  data: expect.arrayContaining([SubscriptionSchema]),
-};
-
-const invoiceSchema: Invoice = {
-  id: expect.any(String),
-  object: expect.any(String),
-  account_country: expect.any(String),
-  account_name: expect.any(String),
-  account_tax_ids: expect.toBeOneOf([expect.toBeArray(), null]),
-  amount_due: expect.any(Number),
-  amount_paid: expect.any(Number),
-  amount_overpaid: expect.any(Number),
-  amount_remaining: expect.any(Number),
-  amount_shipping: expect.any(Number),
-  application: expect.toBeOneOf([expect.any(String), null]),
-  application_fee_amount: expect.toBeOneOf([expect.any(Number), null]),
-  attempt_count: expect.any(Number),
-  attempted: expect.any(Boolean),
-  auto_advance: expect.any(Boolean),
-  automatic_tax: expect.toBeObject(),
-  automatically_finalizes_at: expect.toBeOneOf([expect.any(Number), null]),
-  billing_reason: expect.any(String),
-  charge: expect.any(String),
-  collection_method: expect.any(String),
-  created: expect.any(Number),
-  currency: expect.any(String),
-  custom_fields: expect.toBeOneOf([expect.toBeArray(), null]),
-  customer: expect.any(String),
-  customer_address: expect.toBeOneOf([expect.toBeObject(), null]),
-  customer_email: expect.any(String),
-  customer_name: expect.toBeOneOf([expect.any(String), null]),
-  customer_phone: expect.toBeOneOf([expect.any(String), null]),
-  customer_shipping: expect.toBeOneOf([expect.toBeObject, null]),
-  customer_tax_exempt: expect.any(String),
-  customer_tax_ids: expect.toBeOneOf([expect.toBeArray(), null]),
-  default_payment_method: expect.toBeOneOf([expect.any(String), null]),
-  default_source: expect.toBeOneOf([expect.any(String), null]),
-  default_tax_rates: expect.toBeOneOf([expect.toBeArray(), null]),
-  description: expect.toBeOneOf([expect.any(String), null]),
-  discount: expect.toBeOneOf([expect.toBeObject(), null]),
-  discounts: expect.toBeOneOf([expect.toBeArray(), null]),
-  due_date: expect.toBeOneOf([expect.any(Number), null]),
-  effective_at: expect.any(Number),
-  ending_balance: expect.any(Number),
-  footer: expect.toBeOneOf([expect.any(String), null]),
-  from_invoice: expect.toBeOneOf([expect.toBeObject(), null]),
-  hosted_invoice_url: expect.any(String),
-  invoice_pdf: expect.any(String),
-  issuer: expect.toBeObject(),
-  last_finalization_error: expect.toBeOneOf([expect.toBeObject(), null]),
-  latest_revision: expect.toBeOneOf([expect.any(String), null]),
-  lines: expect.toBeObject(),
-  livemode: expect.any(Boolean),
-  metadata: expect.toBeObject(),
-  next_payment_attempt: expect.toBeOneOf([expect.any(Number), null]),
-  number: expect.any(String),
-  on_behalf_of: expect.toBeOneOf([expect.any(String), null]),
-  paid: expect.any(Boolean),
-  paid_out_of_band: expect.any(Boolean),
-  parent: expect.toBeObject(),
-  payment_intent: expect.any(String),
-  payment_settings: expect.toBeObject(),
-  period_end: expect.any(Number),
-  period_start: expect.any(Number),
-  post_payment_credit_notes_amount: expect.any(Number),
-  pre_payment_credit_notes_amount: expect.any(Number),
-  quote: expect.toBeOneOf([expect.any(String), null]),
-  receipt_number: expect.toBeOneOf([expect.any(String), null]),
-  rendering: expect.toBeOneOf([expect.toBeObject(), null]),
-  rendering_options: expect.toBeOneOf([expect.toBeObject(), undefined]),
-  shipping_cost: expect.toBeOneOf([expect.toBeObject(), null]),
-  shipping_details: expect.toBeOneOf([expect.toBeObject(), null]),
-  starting_balance: expect.any(Number),
-  statement_descriptor: expect.toBeOneOf([expect.any(String), null]),
-  status: expect.any(String),
-  status_transitions: expect.toBeObject(),
-  subscription: expect.any(String),
-  subscription_details: expect.toBeObject(),
-  subtotal: expect.any(Number),
-  subtotal_excluding_tax: expect.any(Number),
-  tax: expect.toBeOneOf([expect.any(Number), null]),
-  test_clock: expect.toBeOneOf([expect.any(String), null]),
-  total: expect.any(Number),
-  total_discount_amounts: expect.toBeOneOf([expect.toBeArray(), null]),
-  total_excluding_tax: expect.any(Number),
-  total_pretax_credit_amounts: expect.toBeOneOf([expect.toBeArray(), null]),
-  total_tax_amounts: expect.toBeArray(),
-  total_taxes: expect.toBeArray(),
-  transfer_data: expect.toBeOneOf([expect.toBeObject(), null]),
-  webhooks_delivered_at: expect.toBeOneOf([expect.any(Number), null]),
-};
-
-const stripeInvoiceSchema: PaginatedSubscriptionInvoice = {
-  first: expect.any(String),
-  last: expect.any(String),
-  hasMore: expect.any(Boolean),
-  data: [invoiceSchema],
-};
-
-const stripePaymentMethodSchema = {
-  id: expect.any(String),
-  object: expect.any(String),
-  allow_redisplay: expect.any(String),
-  billing_details: expect.objectContaining({
-    address: {
-      city: expect.toBeOneOf([expect.any(String), null]),
-      country: expect.toBeOneOf([expect.any(String), null]),
-      line1: expect.toBeOneOf([expect.any(String), null]),
-      line2: expect.toBeOneOf([expect.any(String), null]),
-      postal_code: expect.toBeOneOf([expect.any(String), null]),
-      state: expect.toBeOneOf([expect.any(String), null]),
-    },
-    email: expect.toBeOneOf([expect.any(String), null]),
-    name: expect.toBeOneOf([expect.any(String), null]),
-    phone: expect.toBeOneOf([expect.any(String), null]),
-  }),
-  card: expect.objectContaining({
-    brand: expect.any(String),
-    checks: {
-      address_line1_check: expect.toBeOneOf([expect.any(String), null]),
-      address_postal_code_check: expect.toBeOneOf([expect.any(String), null]),
-      cvc_check: expect.any(String),
-    },
-    country: expect.any(String),
-    display_brand: expect.any(String),
-    exp_month: expect.any(Number),
-    exp_year: expect.any(Number),
-    fingerprint: expect.any(String),
-    funding: expect.any(String),
-    generated_from: expect.toBeOneOf([expect.any(String), null]),
-    last4: expect.any(String),
-    networks: expect.objectContaining({
-      available: expect.toBeArray(),
-      preferred: expect.toBeOneOf([expect.any(String), null]),
-    }),
-    three_d_secure_usage: expect.objectContaining({ supported: expect.any(Boolean) }),
-    wallet: expect.toBeOneOf([expect.any(String), null]),
-  }),
-  created: expect.any(Number),
-  customer: expect.any(String),
-  livemode: expect.any(Boolean),
-  metadata: expect.toBeObject(),
-  type: expect.any(String),
-};
 
 const generateTestData = async () => {
   await prismaClient.subscription.create({
